refactor ShoppingItemProps to take a ShoppingItem type
