+++ smart-cart/src/interfaces/ShoppingItem.tsx
@@ -1,0 +1,8 @@
+type ShoppingItem = {
+  name: string,
+  qty: number,
+  isCrossedOut: boolean,
+  id: number
+}
+
+export default ShoppingItem
