+++ smart-cart/src/interfaces/ShoppingItemProps.tsx
@@ -1,6 +1,7 @@
+import ShoppingItem from "./ShoppingItem";
+
 export default interface ShoppingItemProps {
-  name: string,
-  qty: number,
-  isChecked: boolean,
-  id: number
+  index: number,
+  item: ShoppingItem,
+  handleClick: (index:number) => void,
 }
